refactor: move openFiddle function
